Install the app icon to the hicolor theme
Installing the icon in /usr/local/share/icons/pixmap won't work (only
works in /usr/share/icons/pixmaps). This causes permissions issues. The
current standard way, according to the spec, is:

In order to have a place for third party applications to install their
icons there should always exist a theme called "hicolor". Implementations
are required to look in the "hicolor" theme if an icon was not found in
the current theme.
http://standards.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html